separate auth button widget
