+++ lib/ui/widgets/auth/auth_widget.dart
@@ -40,43 +40,7 @@ class AuthWidget extends StatelessWidget {
 
 
 
-class _AuthButtonWidget extends StatelessWidget {
-  const _AuthButtonWidget({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    final cubit = context.watch<AuthViewCubit>();
-    final authDataStorage = context.read<AuthDataStorage>();
-    final canStartAuth = cubit.state is AuthViewCubitFormFillInProgressState || cubit.state is AuthViewCubitErrorState;
-    const color = Color(0xFF01B4E4);
-    final onPressed =
-        canStartAuth ? () => cubit.auth(login: authDataStorage.login, password: authDataStorage.password) : null;
-    final child = cubit.state is AuthViewCubitAuthProgressState
-        ? const SizedBox(
-            width: 15,
-            height: 15,
-            child: CircularProgressIndicator(strokeWidth: 2),
-          )
-        : const Text('Login');
-    return ElevatedButton(
-      onPressed: onPressed,
-      style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all(color),
-        foregroundColor: MaterialStateProperty.all(Colors.white),
-        textStyle: MaterialStateProperty.all(
-          const TextStyle(fontSize: 16, fontWeight: FontWeight.w700),
-        ),
-        padding: MaterialStateProperty.all(
-          const EdgeInsets.symmetric(
-            horizontal: 15,
-            vertical: 8,
-          ),
-        ),
-      ),
-      child: child,
-    );
-  }
-}
 
 class _ErrorMessageWidget extends StatelessWidget {
   const _ErrorMessageWidget({Key? key}) : super(key: key);
